Refactor SQLParserFactory to adapt SQL parser and Dist SQL parser
